requirements.txt
astroid==2.2.0
atomicwrites==1.3.0
attrs==18.2.0
autoenv==1.0.0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.9.0
isort==4.3.9
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==6.0.0
pluggy==0.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
PyJWT==1.7.1
pylint==2.3.0
pytest==4.3.0
python-dotenv==0.10.1
six==1.12.0
typed-ast==1.2.0
Werkzeug==0.14.1
wrapt==1.11.1
